Revert "Questioncomponent (#8)"
This reverts commit e1a73f9bbbe301844e5d99d998246e938fa96e2a.
