+++ .gitignore
@@ -34,6 +34,3 @@ cypress/cucumber-json
 
 # IDE
 .idea
-
-# oldfiles
-.oldIndexForInspo.js
