+++ src/components/InputLabel/Button.stories.js
@@ -1,0 +1,12 @@
+import InputLabel from './';
+
+export function Default() {
+	return <InputLabel>Button</InputLabel>;
+}
+
+const story = {
+	title: 'InputLabel',
+	component: InputLabel,
+};
+
+export default story;
