+++ src/components/InputLabel/index.js
@@ -1,0 +1,3 @@
+export default function InputLabel({children}) {
+	return <label>{children}</label>;
+}
