--- src/components/QuestionComponent/QuestionComponent.stories.js
@@ -1,12 +1,0 @@
-import QuestionComponent from '.';
-
-export function Default() {
-	return <QuestionComponent>QuestionComponent</QuestionComponent>;
-}
-
-const story = {
-	title: 'QuestionComponent',
-	component: QuestionComponent,
-};
-
-export default story;
