--- src/components/QuestionComponent/index.js
@@ -1,5 +1,0 @@
-import StyledQuestionComponent from './styled';
-
-export default function QuestionComponent({children}) {
-	return <StyledQuestionComponent>{children}</StyledQuestionComponent>;
-}
