+++ src/pages/index.js
@@ -1,15 +1,24 @@
 import {nanoid} from 'nanoid';
+import {useState} from 'react';
 import {Helmet} from 'react-helmet';
 
+import Button from '../components/Button';
 import Layout from '../components/Layout';
-import QuestionComponent from '../components/QuestionComponent';
+import useFetch from '../hooks/useFetch';
+import useStore from '../hooks/useStore';
 
 export default function HomePage() {
-	const questionArray = [
-		{id: nanoid(), questionText: 'What is export default?'},
-		{id: nanoid(), questionText: 'What does clean code means?'},
-		{id: nanoid(), questionText: 'Why do we have to define keys?'},
-	];
+	// Data
+	const {data, loading, error} = useFetch('/api/hello');
+
+	// Local state
+	const [id, setId] = useState(null);
+
+	// Global state ZUSTAND
+	const counter = useStore(state => state.counter);
+	const decrementCounter = useStore(state => state.decrementCounter);
+	const incrementCounter = useStore(state => state.incrementCounter);
+	const setCounter = useStore(state => state.setCounter);
 
 	return (
 		<Layout>
@@ -17,12 +26,51 @@ export default function HomePage() {
 				<title key="title">My Project</title>
 				<meta key="description" name="description" content="This is my project" />
 			</Helmet>
-			{questionArray.map(question => {
-				console.log(question.id);
-				return (
-					<QuestionComponent key={question.id}>{question.questionText}</QuestionComponent>
-				);
-			})}
+			<h1>Home</h1>
+			{loading && <div>Loading...</div>}
+			{error && <div>{error.message}</div>}
+			{data && (
+				<pre>
+					<code>{JSON.stringify(data, null, 4)}</code>
+				</pre>
+			)}
+			<section>
+				<Button
+					aria-label="decrement"
+					onClick={() => {
+						decrementCounter();
+					}}
+				>
+					-
+				</Button>
+				<input
+					value={`${counter}`}
+					size={2}
+					onChange={event => {
+						setCounter(Number.parseInt(event.target.value, 10));
+					}}
+				/>
+				<Button
+					aria-label="increment"
+					onClick={() => {
+						incrementCounter();
+					}}
+				>
+					+
+				</Button>
+			</section>
+			<br />
+			<section>
+				<Button
+					onClick={() => {
+						setId(nanoid());
+					}}
+				>
+					Generate ID
+				</Button>
+				<br />
+				<div>Id: {id}</div>
+			</section>
 		</Layout>
 	);
 }
